+++ resources/release-notes.md
@@ -4,6 +4,11 @@
 * Fix VimRSetCharacterspacing: GH-1014
 * Add basic file handling to the filebrowser: GH-1005
 * Provide neovim original commandline interface: GH-1003
+* Dependencies updates:
+  - ReactiveX/RxSwift@6.6.0
+  - sparkle-project/Sparkle@2.5.0
+  - apple/swift-collections@1.0.5
+  - Quick/Nimble@13.0.0
 
 # 0.44.0-20230103.174333